refactor: update import paths for footer components and add new styles
